+++ dcicutils/misc_utils.py
@@ -927,6 +927,38 @@ def override_dict(d, **overrides):
             d[k] = v
 
 
+@contextlib.contextmanager
+def local_attrs(obj, **kwargs):
+    """
+    This binds the named attributes of the given object.
+    This is only allowed for an object that directly owns the indicated attributes.
+
+    """
+    keys = kwargs.keys()
+    for key in keys:
+        if key not in obj.__dict__:
+            # This works only for objects that directly have the indicated property.
+            # That happens for
+            #  (a) an instance where its instance variables are in keys.
+            #  (b) an uninstantiated class where its class variables (but not inherited class variables) are in keys.
+            # So the error happens for these cases:
+            #  (c) an instance where any of the keys come from its class instead of the instance itself
+            #  (d) an uninstantiated class being used for keys that are inherited rather than direct class variables
+            raise ValueError("%s inherits property %s. Treating it as dynamic could affect other objects."
+                             % (obj, key))
+    saved = {
+        key: getattr(obj, key)
+        for key in keys
+    }
+    try:
+        for key in keys:
+            setattr(obj, key, kwargs[key])
+        yield
+    finally:
+        for key in keys:
+            setattr(obj, key, saved[key])
+
+
 def check_true(test_value, message, error_class=None):
     """
     If the first argument does not evaluate to a true value, an error is raised.
